ContentsWithHelp :
- ContentsWithHelps devient HelpContent
- Renvoi vers la ZEP en commentaires supprimé
- Refactorisation des strings en dur dans le if/elif en important le
module zds.tutorialv2.models afin de pouvoir utiliser les valeurs de
TYPE_CHOICES à la place des str 'TUTORIAL' et 'ARTICLE' codées en durs
- Pluralisation du nom de la class HelpContent via la classe META, qui
devient donc : help-contents au pluriel
