+++ zds/tutorialv2/models/models_database.py
@@ -34,7 +34,7 @@ from zds.utils import get_current_user
 from zds.utils.models import SubCategory, Licence, EditorialHelp, Comment, Tag
 from zds.utils.tutorials import get_blob
 from zds.tutorialv2.models import TYPE_CHOICES, STATUS_CHOICES
-from zds.tutorialv2.models.models_versioned import NotAPublicVersion
+from zds.tutorialv2.models.models_versioned import NotAPublicVersion, EditorialHelp
 from zds.tutorialv2.managers import PublishedContentManager, PublishableContentManager
 import logging
